Cameras and input
Progressed part 3 to 30 minutes in 3-4 hours. Now that's progress!
Anyway, I refactored most of the code in similar way to the code in the video, except I used operator overloading in many cases. Added camera control and input capture(still needs work). Added the screenshots from the videos to the project folder